Update README.md
A sorting visualizer is an interactive tool or software that graphically demonstrates how different sorting algorithms work by visually sorting a collection of elements. It allows users to see the step-by-step process of sorting algorithms and how they manipulate data in real-time. This tool is particularly useful for learning and understanding the inner workings of sorting algorithms.